+++ README.md
@@ -1,37 +1,12 @@
-# **Desafio - Data Analytics**
+# General
 
-Neste desafio buscamos avaliar sua habilidade em análise de dados, utilizando SQL, Python, e storytelling. O desafio consiste em realizar uma análise exploratória dos dados fornecidos, que compreendem informações sobre alunos e escolas na cidade de São Paulo entre 2010 e 2023. Os conjuntos de dados estão disponíveis na pasta **Data**, incluindo arquivos CSV das escolas e do perfil dos educandos.
+Este projeto foi dividido em 4 partes, seguindo um racional para fluxo de dados. Essas partes são: Ingestão, Transformação, Loading e Analytics. 
 
-## **Dados Disponíveis:**
+As etapas de Ingestão, Transformação abrangem as questões de 1 a 4 do desafio e são realizadas pelos arquivos: src/1_extraction.py, notebooks/01_datacleaniing.ipynb. Já as etapas Loading e Analytics são realizadas pelos arquivos src/3_loading.py e através da criação do dashboard disposto na pasta BI, além do link abaixo. 
 
-O banco de dados consiste em mais de +2 milhões de registros, contendo informações detalhadas sobre alunos e escolas. Os atributos de cada tabela estão definidos nos respectivos dicionários de dados.
+link Bi: https://app.powerbi.com/view?r=eyJrIjoiOTEyOTIxZDktYzIxYi00MzdmLThlMGMtODQ1YTAzN2IzYjYxIiwidCI6ImRmNzFmNmJiLWUzY2MtNGY1Yi1iNTMyLTc5ZGUyNjFiNTFhMiJ9
 
-1. 🎲 Escolas (*Diretório com arquivos .csv*)
-2. 🎲 Perfil dos educandos (*Diretório com arquivos .csv*)
-3. ℹ️ Test.md (*Arquivo com as questões do teste)
-4. ℹ️ README.md (*Arquivo informações*)
+# Etapa 1 - 4: Resposta das questôes 1 a 4 
 
-## **Instruções:**
+O primeiro problema a ser resolvido foi extrair o arquivo zipado e loading no sqlite. Isso está disposto no arquivo src/1_extraction.py
 
-Seu objetivo é desenvolver um script ou programa em SQL/Python para resolver as questões propostas no arquivo **Test.md**. Você deve processar todos os arquivos de dados fornecidos, realizando as transformações e limpezas necessárias para garantir que os dados possam ser utilizados para extrair insights relevantes.
-
-Para facilitar a avaliação, siga estas instruções:
-
-- [ ] Faça o upload dos arquivos CSV para um banco SQLite.
-- [ ] Respeite a tipagem dos dados e o nome das colunas ao criar as tabelas no banco de dados.
-- [ ] Desenvolva consultas SQL para responder às perguntas do teste.
-- [ ] Organize seu código em pastas separadas, mantendo a estrutura de pastas fornecida.
-- [ ] Documente sua solução usando comentários ou arquivos README para explicar sua estratégia e o procedimento adotado.
-- [ ] Prepare os commits em branchs separados usando o padrão de nomeação: nome + sobrenome.
-
-Depois de todas as análises e codificação serem feitas, crie uma solicitação de pull (PR) neste repositório.
-
-## **Motivação:**
-
-Este teste avalia sua capacidade de resolver problemas de análise de dados, utilizando habilidades em programação. A capacidade de comunicar seus resultados através de storytelling também será considerada na avaliação.
-
-## **Envio da Solução:**
-
-Para enviar sua solução, faça um fork deste repositório e crie um branch com seu nome e sobrenome. Organize os scripts e código em pastas separadas, mantendo a estrutura original dos arquivos fornecidos. Certifique-se de documentar sua solução para facilitar a compreensão do avaliador.
-
-Certifique-se de completar todas as etapas do desafio e seguir as instruções para enviar sua solução corretamente. Estamos ansiosos para ver sua solução e entender sua abordagem para resolver o problema!
