+++ bi/dashboard_escola.Report/definition.pbir
@@ -1,0 +1,9 @@
+{
+  "version": "4.0",
+  "datasetReference": {
+    "byPath": {
+      "path": "../dashboard_escola.SemanticModel"
+    },
+    "byConnection": null
+  }
+}
