+++ bi/dashboard_escola.SemanticModel/.pbi/editorSettings.json
@@ -1,0 +1,8 @@
+{
+  "version": "1.0",
+  "autodetectRelationships": true,
+  "parallelQueryLoading": true,
+  "typeDetectionEnabled": true,
+  "relationshipImportEnabled": true,
+  "shouldNotifyUserOfNameConflictResolution": true
+}
