+++ bi/dashboard_escola.SemanticModel/definition.pbism
@@ -1,0 +1,4 @@
+{
+  "version": "4.0",
+  "settings": {}
+}
