+++ bi/dashboard_escola.SemanticModel/diagramLayout.json
@@ -1,0 +1,61 @@
+{
+  "version": "1.1.0",
+  "diagrams": [
+    {
+      "ordinal": 0,
+      "scrollPosition": {
+        "x": 198,
+        "y": 0
+      },
+      "nodes": [
+        {
+          "location": {
+            "x": 0,
+            "y": 0
+          },
+          "nodeIndex": "dataset",
+          "nodeLineageTag": "f1f96386-6a55-4d5b-ad76-617b2d5ee70a",
+          "size": {
+            "height": 300,
+            "width": 234
+          },
+          "zIndex": 0
+        },
+        {
+          "location": {
+            "x": 373,
+            "y": 0
+          },
+          "nodeIndex": "dim_date",
+          "nodeLineageTag": "050c03a2-cbb0-4bf3-956c-f7443d7860ac",
+          "size": {
+            "height": 272,
+            "width": 234
+          },
+          "zIndex": 0
+        },
+        {
+          "location": {
+            "x": 639,
+            "y": 0
+          },
+          "nodeIndex": "Metrics",
+          "nodeLineageTag": "6a30672c-cbe6-4b64-8140-c3179bf90bd0",
+          "size": {
+            "height": 104,
+            "width": 234
+          },
+          "zIndex": 0
+        }
+      ],
+      "name": "Todas as tabelas",
+      "zoomValue": 100,
+      "pinKeyFieldsToTop": false,
+      "showExtraHeaderInfo": false,
+      "hideKeyFieldsWhenCollapsed": false,
+      "tablesLocked": false
+    }
+  ],
+  "selectedDiagram": "Todas as tabelas",
+  "defaultDiagram": "Todas as tabelas"
+}
